moved utility functions to new file, fixed assembler.lua xpcall handler improperly handling raw errors, parser validator WIP
